Move system health event source to seperate file
