+++ src/middlewared/middlewared/plugins/system/health.py
@@ -1,0 +1,80 @@
+import psutil
+import time
+
+from middlewared.event import EventSource
+from middlewared.utils import run, start_daemon_thread
+
+
+CACHE_POOLS_STATUSES = 'system.system_health_pools'
+
+
+class SystemHealthEventSource(EventSource):
+
+    """
+    Notifies of current system health which include statistics about consumption of memory and CPU, pools and
+    if updates are available. An integer `delay` argument can be specified to determine the delay
+    on when the periodic event should be generated.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._check_update = None
+        start_daemon_thread(target=self.check_update)
+
+    def check_update(self):
+        while not self._cancel_sync.is_set():
+            try:
+                self._check_update = self.middleware.call_sync('update.check_available')['status']
+            except Exception:
+                self.middleware.logger.warn(
+                    'Failed to check available update for system.health event', exc_info=True,
+                )
+            finally:
+                self._cancel_sync.wait(timeout=60 * 60 * 24)
+
+    def pools_statuses(self):
+        return {
+            p['name']: {'status': p['status']}
+            for p in self.middleware.call_sync('pool.query')
+        }
+
+    def run_sync(self):
+
+        try:
+            if self.arg:
+                delay = int(self.arg)
+            else:
+                delay = 10
+        except ValueError:
+            return
+
+        # Delay too slow
+        if delay < 5:
+            return
+
+        cp_time = psutil.cpu_times()
+        cp_old = cp_time
+
+        while not self._cancel_sync.is_set():
+            time.sleep(delay)
+
+            cp_time = psutil.cpu_times()
+            cp_diff = type(cp_time)(*map(lambda x: x[0] - x[1], zip(cp_time, cp_old)))
+            cp_old = cp_time
+
+            cpu_percent = round(((sum(cp_diff) - cp_diff.idle) / sum(cp_diff)) * 100, 2)
+
+            pools = self.middleware.call_sync(
+                'cache.get_or_put',
+                CACHE_POOLS_STATUSES,
+                1800,
+                self.pools_statuses,
+            )
+
+            self.send_event('ADDED', fields={
+                'cpu_percent': cpu_percent,
+                'memory': psutil.virtual_memory()._asdict(),
+                'pools': pools,
+                'update': self._check_update,
+            })
+
